requirements.txt
python-novaclient
requests
paramiko
selenium
pytest
flake8
unittest2
